% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_games.R
\name{analyse_multiple_players}
\alias{analyse_multiple_players}
\title{Extract Multiple Player's Game Data}
\usage{
analyse_multiple_players(usernames)
}
\arguments{
\item{usernames}{A list of player usernames from chess.com}
}
\description{
\code{analyse_multiple_players} returns a dataframe of game data for a list of usernames
}
\details{
This function will take in a list of player usernames and return
a dataframe of game metadata
}
