% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bits_to_square.R
\name{bits_to_square}
\alias{bits_to_square}
\title{Find out which square is represented in a bitboard}
\usage{
bits_to_square(bitboard, split = FALSE)
}
\arguments{
\item{bitboard}{A bitboard representation, which is a 64 binary vector.}

\item{split}{If true, also returns the square's rank and file.}
}
\value{
A list with the square associated with the bitboard.
}
\description{
Given a bitboard representation, find which square a piece is
occupying.
}
\author{
Eduardo Kapp
}
