% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\docType{class}
\name{board-class}
\alias{board-class}
\alias{board}
\title{Board Class}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Board Class
}
\examples{
add(1, 1)
add(100) sdfs
}
