% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chessR.R
\docType{package}
\name{chessR}
\alias{chessR}
\alias{chessR-package}
\title{chessR: A package for extracting and analysing chess game data}
\description{
This package is designed to aid in the extraction and analysis of game data from the popular chess.com
}
\author{
\strong{Maintainer}: Jason Zivkovic \email{jase.ziv83@gmail.com}

}
