% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{clean_pgn}
\alias{clean_pgn}
\title{clean each pgn string and separate its columns}
\usage{
clean_pgn(t)
}
\arguments{
\item{t}{tibble coming from convert_to_tibble}
}
\description{
clean each pgn string and separate its columns
}
\keyword{internal}
