% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{convert_to_tibble}
\alias{convert_to_tibble}
\title{Extract rules, time class and pgn of each of the games in a list and concentrate them in a tibble}
\usage{
convert_to_tibble(games_list)
}
\arguments{
\item{games_list}{the list of games to extract information from}
}
\description{
Extract rules, time class and pgn of each of the games in a list and concentrate them in a tibble
}
\keyword{internal}
