% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_empty_board.R
\name{create_empty_board}
\alias{create_empty_board}
\title{Define Empty Board}
\usage{
create_empty_board()
}
\value{
A 12 by 64 numeric matrix filled with zeros.
}
\description{
A chess board is represented as a 12 by 64 binary matrix. Each
row represents a piece type with a color. There are 6 piece types, so
to represent whites and blacks, we use 12 rows. Each row has 64 columns,
so that every slot in the 8 by 8 chessboard is represented.
}
\author{
Eduardo Kapp
}
