% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBoard.R
\name{drawBoard}
\alias{drawBoard}
\title{createBoard}
\usage{
drawBoard(numberCol, numberRow)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
createBoard
}
\examples{
add(1, 1)
add(100) sdfs
}
