% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{empty_str_remove}
\alias{empty_str_remove}
\title{Check if a string is non empty to reduce it.}
\usage{
empty_str_remove(string, pattern)
}
\description{
Check if a string is non empty to reduce it.
}
\keyword{internal}
