% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moves.R
\name{extract_moves}
\alias{extract_moves}
\title{Extract moves from a game as a data.frame}
\usage{
extract_moves(moves_string)
}
\arguments{
\item{moves_string}{string containing moves built by \code{chessR} (e.g. from \url{chess.com})}
}
\value{
cleaned moves as a data.frame
}
\description{
Extract moves from a game as a data.frame
}
