% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moves.R
\name{extract_moves_as_game}
\alias{extract_moves_as_game}
\title{Extract moves and create \code{chess} game}
\usage{
extract_moves_as_game(game)
}
\arguments{
\item{game}{a single row of a \code{data.frame} provided by \code{chessR} containing move information}
}
\value{
a \code{\link[chess:game]{chess::game()}} game object
}
\description{
Extract moves and create \code{chess} game
}
