% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{extract_pgn}
\alias{extract_pgn}
\title{Parse and extract game metadata}
\usage{
extract_pgn(x)

extract_pgn(x)
}
\arguments{
\item{x}{list of pgn games}
}
\description{
@keywords internal
}
\keyword{internal}
