% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{extract_time_class}
\alias{extract_time_class}
\title{Extract the time class of each game (ie blitz, bullet, daily, etc)}
\usage{
extract_time_class(x)

extract_time_class(x)
}
\description{
@keywords internal
}
\keyword{internal}
