% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_knight_attacks.R
\name{find_knight_attacks}
\alias{find_knight_attacks}
\title{Find set of knight attacks}
\usage{
find_knight_attacks(board, whites, square)
}
\arguments{
\item{board}{A 12 by 64 binary matrix representing the game current status.}

\item{whites}{If TRUE, the piece is a white piece, else blacks.}

\item{square}{A character representing a square ("a1" to "h8").}
}
\value{
A bitboard with all attacked squared by the Knight.
}
\description{
Given a board state, a side and the knight position, return
every attacked squares by that knight.
}
\author{
Eduardo Kapp
}
