% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_queen_attacks.R
\name{find_queen_attacks}
\alias{find_queen_attacks}
\title{Find set of queen attacks}
\usage{
find_queen_attacks(board, whites, square)
}
\arguments{
\item{board}{A 12 by 64 binary matrix representing the game current status.}

\item{whites}{If TRUE, the piece is a white piece, else blacks.}

\item{square}{A character representing a square ("a1" to "h8").}
}
\value{
A bitboard with all attacked squared by the Queen.
}
\description{
Given a board state, a side and the queen position, return
every attacked squares by that queen
}
\author{
Eduardo Kapp
}
