% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game.R
\docType{class}
\name{game-class}
\alias{game-class}
\alias{game}
\title{Game Class}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Game Class
}
\examples{
add(1, 1)
add(100)
}
