% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_games.R
\name{get_each_player}
\alias{get_each_player}
\title{Extract Single Player's Game Data}
\usage{
get_each_player(username)
}
\arguments{
\item{username}{A string value of a player's name}
}
\description{
\code{analyse_player_games} returns a dataframe of all of the games played by one player
}
\details{
This function will take in a single player's username and return the
 data on all the games they have played on chess.com
}
