% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{get_each_player_chessdotcom}
\alias{get_each_player_chessdotcom}
\title{Get the raw json data for a player's chess.com data as a tibble}
\usage{
get_each_player_chessdotcom(username, year_month)
}
\arguments{
\item{year_month}{the desired year and month in format yyyymm,}

\item{user}{the desired username,}
}
\description{
Get the raw json data for a player's chess.com data as a tibble
}
\keyword{internal}
