% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_game_data.R
\name{get_each_raw}
\alias{get_each_raw}
\title{Get Single Player Raw chess.com Game Data}
\usage{
get_each_raw(username)
}
\arguments{
\item{username}{A valid unsername from chess.com}
}
\description{
This function returns the raw json data for a player's
chess.com data as a data frame
}
