% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_game_data.R
\name{get_game_data}
\alias{get_game_data}
\title{Extract Chess Game Data}
\usage{
get_game_data(usernames)
}
\arguments{
\item{usernames}{A character vector of player usernames from chess.com}
}
\description{
\code{get_chess_data} returns a dataframe of game data for either a single user
or a list of usernames
}
\details{
This function will take in a list of player usernames and return
a dataframe of game metadata
}
