% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{get_game_urls}
\alias{get_game_urls}
\title{Get a list of all year/months the player(s) has played on chess.com}
\usage{
get_game_urls(username)
}
\arguments{
\item{username}{the desired player's nickname}
}
\description{
Get a list of all year/months the player(s) has played on chess.com
}
\keyword{internal}
