% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{get_games}
\alias{get_games}
\title{Parse the list of game urls and extract a json blob}
\usage{
get_games(y)

get_games(y)
}
\arguments{
\item{y}{list of game urls}

\item{dates_list}{a list of game urls by date}
}
\description{
Parse the list of game urls and extract a json blob

Parse the list of game urls and extract a json blob (for get_game_data)
}
\keyword{internal}
