% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_occupied_squares.R
\name{get_occupied_squares}
\alias{get_occupied_squares}
\title{Get Occupied Squares}
\usage{
get_occupied_squares(board)
}
\arguments{
\item{board}{A 12 by 64 numeric matrix.}
}
\value{
A 1 by 64 bitboard where every 1 represent an occupied slot.
}
\description{
Return 1 for every square that has a piece in it.
}
