% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_chessdotcom.R
\name{get_raw_chessdotcom}
\alias{get_raw_chessdotcom}
\title{Get Raw chess.com Game Data}
\usage{
get_raw_chessdotcom(usernames, year_month = NA_integer_)
}
\arguments{
\item{usernames}{A vector of a valid unsername or usernames from chess.com}

\item{year_month}{An integer of YYYYMM}
}
\description{
This function returns the raw json data for a player's or list of players'
chess.com data as a data frame, for all or select months played
}
