% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_game_data.R
\name{get_raw_game_data}
\alias{get_raw_game_data}
\title{Get Raw chess.com Game Data}
\usage{
get_raw_game_data(usernames)
}
\arguments{
\item{usernames}{A vector of a valid unsername or usernames from chess.com}
}
\description{
This function returns the raw json data for a player's or list of players'
chess.com data as a date frame
}
