% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_lichess.R
\name{get_raw_lichess}
\alias{get_raw_lichess}
\title{Get Raw Lichess Game Data}
\usage{
get_raw_lichess(player_names)
}
\arguments{
\item{player_names}{A vector of a valid username or usernames from Lichess}
}
\value{
a dataframe of lichess data
}
\description{
This function returns the raw json data for a player's or list of players'
Lichess data as a data frame
}
\examples{
\dontrun{
georges_data <- get_games_lichess(player_names = "Georges")
}
}
