% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top50_leaderboard.R
\name{get_top50_leaderboard}
\alias{get_top50_leaderboard}
\title{Get Top 50 on Leaderboards}
\usage{
get_top50_leaderboard(game_type)
}
\arguments{
\item{game_type}{A valid chess.com game type to return the leaderboard for}
}
\description{
This function takes in one parameter, the game_type, and returns a
data frame of the top 50 players.
}
\details{
The leaderboard options (games) include:

\emph{"daily"}, \emph{"daily960}", \emph{"live_rapid"},
\emph{"live_blitz"}, \emph{"live_bullet"}, \emph{"live_bughouse"},
\emph{"live_blitz960"}, \emph{"live_threecheck" }, \emph{"live_crazyhouse"},
\emph{"live_kingofthehill"}, \emph{"lessons"}, \emph{"tactics"}
}
