% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{get_winner}
\alias{get_winner}
\title{Return the game winner}
\usage{
get_winner(result_column, white, black)
}
\arguments{
\item{result_column}{A character vector in the extracted data frame called 'Result'}

\item{white}{A character vector in the extracted data frame called 'White' for the player on white}

\item{black}{A character vector in the extracted data frame called 'Black' for the player on black}
}
\value{
A character vector of the game ending for each game
}
\description{
This function returns a character vector of the usernames of the game winners
}
\examples{
\dontrun{
get_winner(df$Result, df$White, df$Black)
}

}
