% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEMP.R
\name{lichess_clock_move_time}
\alias{lichess_clock_move_time}
\title{Get Time Information from Lichess Game Data}
\usage{
lichess_clock_move_time(games_list)
}
\arguments{
\item{games_list}{A data frame of lichess data which can be generated from chessR::get_raw_lichess("username")}
}
\value{
a data frame of lichess data with move time, clock time, and move numbers
}
\description{
This function returns a data frame
of Lichess data with clock and move times
}
\examples{
\dontrun{
lordy_leroy_data <- get_raw_lichess(player_names = "LordyLeroy")
lordy_leroy_data_with_times <- lichess_clock_move_time(games_list = lordy_leroy_data)
}
}
