% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_piece_name.R
\name{lookup_piece_name}
\alias{lookup_piece_name}
\title{Piece lookup table}
\usage{
lookup_piece_name(x)
}
\arguments{
\item{x}{A number from 0 to 12}
}
\value{
A character representing the piece name
}
\description{
A simple lookup function that takes any number from 0 to 12
and maps it to a piece name.
}
