% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moves.R
\name{pgn_to_dataframe}
\alias{pgn_to_dataframe}
\title{Extract moves from a game as a data.frame}
\usage{
pgn_to_dataframe(moves_string)
}
\arguments{
\item{moves_string}{string containing moves built by `chessR` (e.g. from \url{https://www.chess.com/})
or the filename of a PGN file}
}
\value{
cleaned moves as a data.frame
}
\description{
Extract moves from a game as a data.frame
}
