% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moves.R
\name{pgn_to_game}
\alias{pgn_to_game}
\title{Extract moves and create `chess` game}
\usage{
pgn_to_game(game)
}
\arguments{
\item{game}{a single row of a `data.frame` provided by `chessR` containing move information
or the filename of a PGN file}
}
\value{
a [chess::game()] game object
}
\description{
Extract moves and create `chess` game
}
