% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moves.R
\name{plot_game}
\alias{plot_game}
\title{Plot a game}
\usage{
plot_game(game, interactive = TRUE)
}
\arguments{
\item{game}{a [chess::game()] object, likely with moves identified}

\item{interactive}{wait for 'Enter' after each move? Turn off to use in a gif}
}
\value{
`NULL`, (invisibly) - called for the side-effect of plotting
}
\description{
Plot a game
}
\examples{
\dontrun{
hikaru <- chessR:::get_each_player_chessdotcom("hikaru", "202112")
m <- extract_moves_as_game(hikaru[11, ])
plot_game(m)
}
}
