% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moves.R
\name{plot_moves}
\alias{plot_moves}
\title{Plot a game}
\usage{
plot_moves(game, interactive = TRUE, sleep = 1)
}
\arguments{
\item{game}{a [chess::game()] object, likely with moves identified}

\item{interactive}{wait for 'Enter' after each move? Turn off to use in a gif}

\item{sleep}{how long to wait between moves}
}
\value{
`NULL`, (invisibly) - called for the side-effect of plotting
}
\description{
Plot a game
}
