% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{return_num_moves}
\alias{return_num_moves}
\title{Return the number of moves in a game}
\usage{
return_num_moves(moves_string)
}
\arguments{
\item{moves_string}{A character vector of chess Moves data in PGN notation usually called 'Moves' in extracted data}
}
\value{
A numeric vector of the number of moves in each game
}
\description{
This function returns the number of moves played in each game. The function
accepts a vector of chess Moves data in PGN notation, usually called 'Moves'
}
\examples{
\dontrun{
return_num_moves(moves_string = df$Moves)
}

}
