% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square_to_bits.R
\name{square_to_bits}
\alias{square_to_bits}
\title{Generate a bitboard representation of a square}
\usage{
square_to_bits(square)
}
\arguments{
\item{square}{Any square from "a1" to "h8".}
}
\value{
A 64 binary vector.
}
\description{
Given a Square, generate its bitboard integer.
}
\author{
Eduardo Kapp
}
