% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_board.R
\name{start_board}
\alias{start_board}
\title{Set board to initial state}
\usage{
start_board(board)
}
\arguments{
\item{board}{A 12 by 64 numeric matrix.}
}
\description{
Given a board matrix, set it to chess initial state.
}
