% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_chessdotcom.R
\name{get_each_player_chessdotcom}
\alias{get_each_player_chessdotcom}
\title{Get Single Player Raw chess.com Game Data}
\usage{
get_each_player_chessdotcom(username, year_month)
}
\arguments{
\item{username}{A valid unsername from chess.com}

\item{year_month}{An integer of YYYYMM}
}
\description{
This function returns the raw json data for a player's
chess.com data as a data frame
}
