% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{get_game_ending}
\alias{get_game_ending}
\title{Return the game ending}
\usage{
get_game_ending(termination_string, white, black)
}
\arguments{
\item{termination_string}{A character vector in the chess.com extracted data frame called 'Termination'}

\item{white}{A character vector in the chess.com extracted data frame called 'White' for the player on white}

\item{black}{A character vector in the chess.com extracted data frame called 'Black' for the player on black}
}
\value{
A character vector of the game ending for each game
}
\description{
This function returns a character vector of how the game ended from chess.dom.
}
\examples{
\dontrun{
get_game_ending(termination_string = df$Termination, df$White, df$Black)
}

}
