% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_chessdotcom.R
\name{get_raw_chessdotcom}
\alias{get_raw_chessdotcom}
\title{Get Raw chess.com Game Data}
\usage{
get_raw_chessdotcom(usernames, year_month = NA_integer_)
}
\arguments{
\item{usernames}{A vector of a valid username or usernames from chess.com}

\item{year_month}{An integer of YYYYMM}
}
\value{
a dataframe of chessdotcom data
}
\description{
This function returns the raw json data for a player's or list of players'
chess.com data as a data frame, for all or select months played
}
\examples{
\dontrun{
get_raw_chessdotcom(usernames = "JaseZiv")
get_raw_chessdotcom(usernames = "JaseZiv", year_month = c(202112:202201))
get_raw_chessdotcom(usernames = c("JaseZiv", "Smudgy1"), year_month = 202201)
}
}
