% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_board.R
\name{print_board}
\alias{print_board}
\title{Print board to user}
\usage{
print_board(board)
}
\arguments{
\item{board}{A 12 by 64 numeric matrix.}
}
\description{
Given a board matrix, convert it to a user-friendly 8x8 matrix
with every slot represented with piece names.
}
